Implement user existence check in database
Each "name=" component is in 2+7n row. This code read the str after the "name=" in each 2+7n row and check if it is as same as the parameter  "username[]". Therefore, any '\n' except the one at the end of each component is contradictive to the mechanism of this function.